% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{cooladataConnection}
\alias{cooladataConnection}
\title{Cooladata Connection}
\usage{
cooladataConnection(project_id)
}
\arguments{
\item{project_id}{the internal Cooladata ID of the evironment where to execute the query.}
}
\value{
the function returns the connection that takes query string as an argument and return a data frame
}
\description{
This function creates a connection to Cooladata. The connection can be used to run CQL queries.
The function requires Cooladata developer token in your home repository.
The token should be stored in ~/.cooladatatoken
}
\examples{
cooladata <- cooladataConnection(123456)
cooladata("SELECT event_name FROM cooladata WHERE date_range(last 7 days)  limit 10")

}
\keyword{API}
\keyword{Cooladata}
\keyword{R}
\keyword{package}
